requirements.txt
aiohttp==3.9.5
aiohttp-retry==2.8.3
aiosignal==1.3.1
asgiref==3.8.1
attrs==23.2.0
autopep8==2.1.0
certifi==2024.7.4
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==42.0.8
defusedxml==0.7.1
dj-database-url==2.2.0
dj-rest-auth==6.0.0
Django==5.0.7
django-allauth==0.63.6
django-cors-headers==4.4.0
django-polymorphic==3.1.0
django-rest-polymorphic==0.1.10
django-sms==0.7.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.27.2
frozendict==2.4.4
frozenlist==1.4.1
gunicorn==22.0.0
idna==3.7
inflection==0.5.1
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
multidict==6.0.5
nulltype==2.3.1
oauthlib==3.2.2
packaging==24.1
plaid-python==24.0.0
psycopg2-binary==2.9.9
pycodestyle==2.11.1
pycparser==2.22
PyJWT==2.8.0
pyserial==3.5
python-dateutil==2.9.0.post0
python3-openid==3.2.0
PyYAML==6.0.1
referencing==0.35.1
requests==2.32.3
requests-oauthlib==2.0.0
rpds-py==0.19.0
setuptools==70.3.0
six==1.16.0
sms==0.4
snaptrade-python-sdk==11.0.28
sqlparse==0.5.1
stripe==10.3.0
twilio==9.2.3
typing_extensions==4.12.2
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.2
yarl==1.9.4
